foodcartapp/serializers.py
from django.db import transaction
from rest_framework.serializers import ModelSerializer

from .models import Order, OrderItem


class OrderItemSerializer(ModelSerializer):
    class Meta:
        model = OrderItem
        fields = ['product', 'quantity']


class OrderSerializer(ModelSerializer):
    products = OrderItemSerializer(many=True, allow_empty=False, write_only=True)

    class Meta:
        model = Order
        fields = ['products', 'firstname', 'lastname', 'phonenumber', 'address']

    @transaction.atomic
    def create(request, validated_data):
        order = Order.objects.create(
            firstname=validated_data['firstname'],
            lastname=validated_data['lastname'],
            phonenumber=validated_data['phonenumber'],
            address=validated_data['address'],
        )

        order_items = validated_data['products']
        current_order_items = [OrderItem(
            order=order,
            price=product['product'].price,
            **product,
        ) for product in order_items]
        OrderItem.objects.bulk_create(current_order_items)
        return order
